pyproject.toml
[project]
name = "ai-customer-support"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "langchain>=1.2.13",
    "langchain-community>=0.4.1",
    "langchain-ollama>=1.0.1",
    "langchainhub>=0.1.21",
    "openai>=2.30.0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "uvicorn>=0.42.0",
]
